Moved enums to common
