+++ docs/orders.rst
@@ -10,48 +10,30 @@ a rather complex JSON object that describes the desired order. TDA provides some
 <https://developer.tdameritrade.com/content/place-order-samples>`__ to
 illustrate the process and provides a schema in the `place order documentation 
 <https://developer.tdameritrade.com/account-access/apis/post/accounts/
-%7BaccountId%7D/orders-0>`__, but beyond that we're on our own.
+%7BaccountId%7D/orders-0>`__, but beyond that we're on our own. ``tda-api`` aims 
+to be useful to everyone, from users who want to easily place common equities 
+and options trades, to advanced users who want to place complex multi-leg, 
+multi-asset type trades. 
 
-The ``tda.orders`` module provides an incomplete set of helpers for building 
-these order specs. The aim is to make it impossible to build an invalid JSON 
-object. For example, here is how you might use this module to place a market 
-order for ten shares of Apple common stock:
+For users interested in simple trades, ``tda-api`` supports pre-built trade 
+templates that allow fast construction of the most common trades. Advanced users 
+can modify these trades however they like, and can even build trades from 
+scratch.
 
-.. code-block:: python
+This section begins by listing the supported common trades, and then describing 
+the fully-featured order construction system that supports them. Advanced users 
+can skip straight to that section.
 
-  from tda.orders import EquityOrderBuilder, Duration, Session
 
-  builder = EquityOrderBuilder('AAPL', 10)
-  builder.set_instruction(EquityOrderBuilder.Instruction.SELL)
-  builder.set_order_type(EquityOrderBuilder.OrderType.MARKET)
-  builder.set_duration(Duration.DAY)
-  builder.set_session(Session.NORMAL)
+------------------
+Common Order Types
+------------------
 
-  client = ...  # Get a client however you see fit
-  account_id = 12345678
+This section describes how to construct the most common order types. If you feel 
+there's an order missing, you're welcome to `request a new one
+<https://github.com/alexgolec/tda-api/issues>`__.
 
-  resp = client.place_order(account_id, builder.build())
-  assert resp.ok
 
--------------
-Common Values
--------------
-
-.. autoclass:: tda.orders.Duration
-  :members:
-  :undoc-members:
-.. autoclass:: tda.orders.Session
-  :members:
-  :undoc-members:
-
-.. autoexception:: tda.orders.InvalidOrderException
-
--------------
-Equity Orders
--------------
-
-.. autoclass:: tda.orders.EquityOrderBuilder
-  :members:
-  :undoc-members:
-
-  .. automethod:: __init__
+++++++++++++++++++++++++++++++++++++
+Buying and Selling Equities and ETFs
+++++++++++++++++++++++++++++++++++++
