+++ tests/orders/generic_test.py
@@ -1,6 +1,7 @@
 import unittest
 
 from tda.orders.generic import *
+from tda.orders.common import *
 from tests.test_utils import has_diff
